requirements.txt
six
numpy
Pillow
tensorflow
matplotlib
